Remove FeatureScope Remove Singleton scope Remove CoreComponent Reorganise modules
